requirements.txt
annotated-types==0.7.0
anyio==4.8.0
apturl==0.5.2
asgiref==3.5.0
async-timeout==4.0.3
asyncpg==0.30.0
bcrypt==3.2.0
blinker==1.4
Brlapi==0.8.3
certifi==2020.6.20
cffi==1.17.1
chardet==4.0.0
charset-normalizer==3.4.1
click==8.0.3
colorama==0.4.4
command-not-found==0.3
cryptography==44.0.2
cupshelpers==1.0
databases==0.9.0
dbus-python==1.2.18
defer==1.0.6
distro==1.7.0
distro-info==1.1+ubuntu0.2
duplicity==0.8.21
ecdsa==0.19.0
exceptiongroup==1.2.2
fastapi==0.115.11
fasteners==0.14.1
future==0.18.2
greenlet==3.1.1
groq==0.18.0
gyp==0.1
h11==0.13.0
httpcore==1.0.7
httplib2==0.20.2
httpx==0.28.1
idna==3.3
importlib-metadata==4.6.4
iniconfig==2.0.0
jeepney==0.7.1
jsonpatch==1.33
jsonpointer==3.0.0
keyring==23.5.0
langchain==0.3.20
langchain-core==0.3.43
langchain-text-splitters==0.3.6
langsmith==0.3.13
language-selector==0.1
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
lockfile==0.12.2
louis==3.20.0
macaroonbakery==1.3.1
Mako==1.1.3
MarkupSafe==2.0.1
monotonic==1.6
more-itertools==8.10.0
netifaces==0.11.0
numpy==2.2.3
oauthlib==3.2.0
olefile==0.46
orjson==3.10.15
packaging==24.2
paramiko==2.9.3
passlib==1.7.4
pdfminer.six==20240706
pexpect==4.8.0
Pillow==9.0.1
pluggy==1.5.0
protobuf==3.12.4
psycopg2-binary==2.9.10
ptyprocess==0.7.0
pyasn1==0.4.8
pycairo==1.20.1
pycparser==2.22
pycups==2.0.1
pydantic==2.10.6
pydantic_core==2.27.2
PyGObject==3.42.1
PyJWT==2.3.0
pymacaroons==0.13.0
PyNaCl==1.5.0
pyparsing==2.4.7
pyRFC3339==1.1
pytest==8.3.5
pytest-asyncio==0.25.3
python-apt==2.4.0+ubuntu3
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==1.0.1
python-jose==3.4.0
python-multipart==0.0.20
pytz==2022.1
pyxdg==0.27
PyYAML==5.4.1
reportlab==3.6.8
requests==2.25.1
requests-toolbelt==1.0.0
rsa==4.9
SecretStorage==3.3.1
six==1.16.0
sniffio==1.3.1
SQLAlchemy==2.0.38
starlette==0.46.0
systemd-python==234
tenacity==9.0.0
tomli==2.2.1
typing_extensions==4.12.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
usb-creator==0.3.7
uvicorn==0.15.0
wadllib==1.3.6
wsproto==1.0.0
xdg==5
xkit==0.0.0
zipp==1.0.0
zstandard==0.23.0
